@@ -1,5 +1,5 @@
 {
-  "name": "e-learning",
+  "name": "vutemp",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,22 +9,18 @@
     "i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\""
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/vue-fontawesome": "^2.0.6",
+    "@fortawesome/fontawesome-free": "^5.15.4",
     "core-js": "^3.6.5",
-    "sass-loader": "^10",
+    "rm": "^0.1.8",
     "vee-validate": "^4.5.7",
-    "vue": "^2.6.14",
-    "vue-i18n": "^8.26.3",
-    "vue-router": "^3.5.3",
-    "vue-styled-components": "^1.6.0",
-    "vue-template-compiler": "^2.6.14",
-    "vuetify": "^2.4.0"
+    "vue": "^3.0.0",
+    "vue-i18n": "^9.1.0",
+    "vue-router": "^4.0.0-0",
+    "vue3-styled-components": "^1.2.1",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.4",
+    "@intlify/vue-i18n-loader": "^3.0.0",
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-plugin-router": "^4.5.15",
@@ -33,20 +29,12 @@
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^7.0.0",
-    "sass": "~1.32.0",
-    "sass-loader": "^10.0.0",
-    "vue-cli-plugin-i18n": "~2.3.1",
-    "vue-cli-plugin-vuetify": "~2.4.5",
-    "vuetify-loader": "^1.7.0"
+    "vue-cli-plugin-i18n": "~2.3.1"
   },
   "eslintConfig": {
     "root": true,
     "env": {
-      "node": true,
-      "commonjs": true,
-      "browser": true,
-      "es6": true,
-      "jest": true
+      "node": true
     },
     "extends": [
       "plugin:vue/vue3-essential",
@@ -57,6 +45,7 @@
     },
     "rules": {
       "no-unused-vars": 0,
+      "vue/no-unused-vars": 0,
       "vue/no-unused-components": 0
     }
   },
