@@ -1,39 +1,35 @@
-import Vue from "vue";
-import VueRouter from "vue-router";
-
-Vue.use(VueRouter);
+import { createRouter, createWebHistory } from "vue-router";
 
 const routes = [
   {
     path: "/",
-    name: "Welcome",
-    component: () => import("../views/Welcome"),
-  },
-  {
-    path: "/home",
     name: "Home",
-    component: () => import("../views/Home"),
-  },
-  {
-    path: "/login",
-    name: "Login",
-    component: () => import("../views/Login"),
+    component: () => import("@/views/Home"),
   },
   {
     path: "/about",
     name: "About",
-    component: () => import("../views/About"),
+    component: () => import("@/views/About"),
+  },
+  {
+    path: "/login",
+    name: "Login",
+    component: () => import("@/views/Login"),
+  },
+  {
+    path: "/welcome",
+    name: "Welcome",
+    component: () => import("@/views/Welcome"),
   },
   {
     path: "/contact-us",
-    name: "Contact Us",
-    component: () => import("../views/ContactUs"),
+    name: "ContactUs",
+    component: () => import("@/views/ContactUs"),
   },
 ];
 
-const router = new VueRouter({
-  mode: "history",
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
   routes,
 });
 
